Changes to make things faster, use less memory, and bug fixes

+++ utils/cinclude/group_points_gpu.h
@@ -6,12 +6,12 @@ extern "C" {
 #endif
 
 void group_points_kernel_wrapper(int b, int n, int c, int npoints, int nsample,
-				 const float *points, const long *idx,
+				 const float *points, const int *idx,
 				 float *out, cudaStream_t stream);
 
 void group_points_grad_kernel_wrapper(int b, int n, int c, int npoints,
 				      int nsample, const float *grad_out,
-				      const long *idx, float *grad_points,
+				      const int *idx, float *grad_points,
 				      cudaStream_t stream);
 #ifdef __cplusplus
 }
